Refactor focused transducer module into clean, modular architecture
