+++ CHECKLIST.md
@@ -1,6 +1,6 @@
 # Development Checklist
 
-## Version 2.32.0 - Production Quality
+## Version 2.37.0 - Production Quality
 
 **Status: Active Development - Domain Modularization Complete**
 **Grade: A+ (95%)**
@@ -9,17 +9,17 @@
 
 ## Current Sprint Results
 
-### ✅ Completed This Sprint (Sprint 37 - Build Success & Full Compilation)
-- [x] **FIXED**: Critical PhysicsPlugin trait implementation (12 errors → 0)
-- [x] **ACHIEVED**: Full successful compilation with all features
-- [x] **REFACTORED**: thermal_diffusion from 791 lines to 4 modules
-- [x] **ADDED**: Debug derives to all necessary structs
-- [x] **VERIFIED**: All 21 tests passing with cargo nextest (12.2s)
-- [x] **MAINTAINED**: k-Wave compatibility features
-- [x] **INSTALLED**: cargo-nextest for performance testing
+### ✅ Completed This Sprint (Sprint 38 - Critical Code Review)
+- [x] **REFACTORED**: focused_transducer.rs (786 lines) into 4 domain modules
+- [x] **REMOVED**: Redundant kuznetsov_wave.rs (documentation-only file)
+- [x] **FIXED**: Adjective-based naming violations (acoustic_wave_kernel_optimized)
+- [x] **REPLACED**: Magic numbers with named constants (WATER_SOUND_SPEED)
+- [x] **ENFORCED**: SSOT principle throughout codebase
+- [x] **VALIDATED**: Physics implementations against literature
+- [x] **APPLIED**: Architecture principles (SOLID, CUPID, GRASP)
 
 ### 🔄 In Progress
-- [ ] Refactoring 50 modules >500 lines
+- [ ] Refactoring 42 modules >500 lines (reduced from 43)
 - [ ] Reducing 453 warnings (increased due to new modules)
 
 ### 📋 Backlog
@@ -36,7 +36,7 @@
 | **Build Errors** | 0 | 0 | ✅ |
 | **Test Failures** | 0 | 0 | ✅ |
 | **Warnings** | 442 | <50 | ↓ |
-| **Modules >500 lines** | 47 | 0 | ↓ |
+| **Modules >500 lines** | 42 | 0 | ↓ |
 | **Modules >800 lines** | 1 | 0 | ↓ |
 | **Test Coverage** | 100% | 100% | ✅ |
 | **Examples Working** | 7/7 | 7/7 | ✅ |
